Updating  the order aggregate

+++ Core/Entities/OrderAggregate/Order.cs
@@ -6,13 +6,13 @@
         {
         }
         public Order(IReadOnlyList<OrderItem> orderItems, string buyerEmail, Address shipToAddress,
-            DeliveryMethod deliveryMethod, decimal subtotal, string paymentIntentId)
+            DeliveryMethod deliveryMethod, double subtotal, string paymentIntentId)
         {
             BuyerEmail = buyerEmail;
             ShipToAddress = shipToAddress;
             DeliveryMethod = deliveryMethod;
             OrderItems = orderItems;
-            Subtotal = subtotal;
+            Subtotal = subtotal;    
             PaymentIntentId = paymentIntentId;
         }
 
@@ -21,11 +21,11 @@
         public Address ShipToAddress { get; set; }
         public DeliveryMethod DeliveryMethod { get; set; }
         public IReadOnlyList<OrderItem> OrderItems { get; set; }
-        public decimal Subtotal { get; set; }
+        public double Subtotal { get; set; }
         public OrderStatus Status { get; set; } = OrderStatus.Pending;
         public string PaymentIntentId { get; set; }
 
-        public decimal GetTotal()
+        public double GetTotal()
         {
             return Subtotal + DeliveryMethod.Price;
         }
